Refactor: 이미지 barrel 적용
