+++ .gitignore
@@ -1,5 +1,4 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-.env
 # dependencies
 /node_modules
 /.pnp
