+++ public/index.ts
@@ -1,12 +1,42 @@
 export { default as AlertIcon } from './svgs/alertIcon.svg';
+export { default as CameraIcon } from './svgs/camera.svg';
+export { default as CanvasLoadingIcon } from './svgs/canvasLoading.svg';
+export { default as CaretLeftIcon } from './svgs/caretLeft.svg';
+export { default as CaretRightIcon } from './svgs/caretRight.svg';
+export { default as CartIcon } from './svgs/cart.svg';
+export { default as CheckboxCircleIcon } from './svgs/checkboxCircle.svg';
 export { default as ChevronIcon } from './svgs/chevron.svg';
+export { default as CommentIcon } from './svgs/comment.svg';
+export { default as CrossCircleIcon } from './svgs/crossCircle.svg';
+export { default as DeleteIcon } from './svgs/delete.svg';
 export { default as EyeOffIcon } from './svgs/eyeOff.svg';
 export { default as EyeOnIcon } from './svgs/eyeOn.svg';
+export { default as HeartIcon } from './svgs/heart.svg';
 export { default as LogoIcon } from './svgs/logo.svg';
+export { default as PlusIcon } from './svgs/plus.svg';
 export { default as SearchIcon } from './svgs/search.svg';
+export { default as SoundVibrationIcon } from './svgs/soundVibration.svg';
+export { default as StarIcon } from './svgs/star.svg';
+export { default as StepCompleteIcon } from './svgs/stepComplete.svg';
+export { default as ThumbIcon } from './svgs/thumb.svg';
+export { default as UpArrowIcon } from './svgs/upArrow.svg';
 export { default as UserIcon } from './svgs/user.svg';
+export { default as VerticalTripleDotIcon } from './svgs/verticalTripleDot.svg';
 
 export { default as blackSwitchImg } from './images/blackSwitch.jpg';
 export { default as blueSwitchImg } from './images/blueSwitch.jpg';
 export { default as brownSwitchImg } from './images/brownSwitch.jpg';
+export { default as guide1Img } from './images/guide1.png';
+export { default as guide2Img } from './images/guide2.png';
+export { default as guide3Img } from './images/guide3.png';
+export { default as guide4Img } from './images/guide4.png';
+export { default as heroItem1Img } from './images/heroItem1.png';
+export { default as heroItem2Img } from './images/heroItem2.png';
+export { default as heroItem3Img } from './images/heroItem3.png';
+export { default as heroItem4Img } from './images/heroItem4.png';
+export { default as heroItem5Img } from './images/heroItem5.png';
+export { default as heroItem6Img } from './images/heroItem6.png';
+export { default as keydeukProfileImg } from './images/kedeukProfile.png';
+export { default as keydeukImg } from './images/keydeuk.png';
+export { default as myProfileImg } from './images/myProfile.jpeg';
 export { default as redSwitchImg } from './images/redSwitch.jpg';
