+++ src/app/_components/heroItem.ts
@@ -1,10 +1,5 @@
-import heroItem1 from '@/public/images/heroItem1.png';
-import heroItem2 from '@/public/images/heroItem2.png';
-import heroItem3 from '@/public/images/heroItem3.png';
-import heroItem4 from '@/public/images/heroItem4.png';
-import heroItem5 from '@/public/images/heroItem5.png';
-import heroItem6 from '@/public/images/heroItem6.png';
+import { heroItem1Img, heroItem2Img, heroItem3Img, heroItem4Img, heroItem5Img, heroItem6Img } from '@/public/index';
 
-const images = [heroItem1, heroItem2, heroItem3, heroItem4, heroItem5, heroItem6];
+const images = [heroItem1Img, heroItem2Img, heroItem3Img, heroItem4Img, heroItem5Img, heroItem6Img];
 
 export default images;
