+++ src/app/community/_components/AuthorCard.tsx
@@ -1,7 +1,6 @@
 import classNames from 'classnames/bind';
 
-import myProfileImage from '@/public/images/myProfile.jpeg';
-import ShowMoreIcon from '@/public/svgs/verticalTripleDot.svg';
+import { VerticalTripleDotIcon, myProfileImg } from '@/public/index';
 
 import ProfileImage from '@/components/ProfileImage/ProfileImage';
 
@@ -17,13 +16,13 @@ interface AuthorCardProps {
 export default function AuthorCard({ nickname, dateText }: AuthorCardProps) {
   return (
     <div className={cn('container')}>
-      <ProfileImage profileImage={myProfileImage} />
+      <ProfileImage profileImage={myProfileImg} />
       <div className={cn('info-textbox')}>
         <p className={cn('user-name')}>{nickname}</p>
         <p className={cn('sub-info')}>{dateText}</p>
       </div>
       <div className={cn('show-more-icon')}>
-        <ShowMoreIcon />
+        <VerticalTripleDotIcon />
       </div>
     </div>
   );
