+++ src/app/custom-keyboard/_components/option/parts/ColorTag.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames/bind';
 import type { CustomKeyboardKeyTypes } from '@/types/CustomKeyboardTypes';
 import { Color } from '@react-three/fiber';
 
-import DeleteTagIcon from '@/public/svgs/delete.svg';
+import { DeleteIcon } from '@/public/index';
 
 import styles from './ColorTag.module.scss';
 
@@ -20,7 +20,7 @@ export default function ColorTag({ keyCap, color, onDelete }: ColorTagProps) {
   return (
     <div className={cn('wrapper')}>
       <div className={cn('color-text')}>{`${keyCap} : ${color.toString().toUpperCase()}`}</div>
-      <DeleteTagIcon width={20} height={20} onClick={onDelete} className={cn('delete-icon')} />
+      <DeleteIcon width={20} height={20} onClick={onDelete} className={cn('delete-icon')} />
     </div>
   );
 }
