+++ src/app/layout.tsx
@@ -9,7 +9,7 @@ import { Providers } from './providers';
 import '@/styles/reset.css';
 
 export const metadata: Metadata = {
-  title: '키보드 득템 :: KeyDuek',
+  title: '키보드 득템 :: KeyDeuk',
   description: 'Generated by create next app',
 };
